Updated 'food product analog'
Under discussion:
Should it be 'imitate a given food' or substitute a given food'?
Should 'food substitute', 'food alternative' be the synonyms?
Another definition:  A manufactured substance that is used in place of a traditional food or a food ingredient.
Def Source: https://archives.gadoe.org › GACTE_food_analog6PDF