`poultry or game bird` subclasses -> siblings
Moved all `poultry or game bird` FOODON:03411563 subclasses up a level so they are now sibling classes to it and direct subclasses of `avian animal`.

`hen` FOODON:00003282 had an equivalence axiom that included `poultry or game bird` so I subbed in `poultry` instead. Not sure if this is the right call.