deprecated: FOODON:00003907
Original term accidently created with an "alternative label" instead of a "label" and thus wasn't populating search results properly and a new term was inadvertently made to replace it. Deprecated this version under the assumption it wasn't adopted without a label.

Additionally: moved the `example of usage` to the `term replaced by` poultry.