Major release
Main change, among many, is that over 1,700 plant terms and a few animal terms were un-deprecated, to correct a design mistake that didn't separate NCBITaxon references from references to a whole plant or animal of the same taxon.  NCBITaxon doesn't commit to a whole rather than a part of an organism.